+++ Project_SDL_Part1_base/Project_SDL1.h
@@ -18,4 +18,4 @@ constexpr unsigned frame_boundary = 100;
 // Helper function to initialize SDL
 void init();
 SDL_Surface* load_surface_for(const std::string& path, SDL_Surface* window_surface_ptr);
-void constrained_linear_move_(unsigned int& x, unsigned int& y, int& vx, int& vy);
+void constrained_linear_move_(double& x, double& y, double& vx, double& vy);
